Fixed bug relating inferring segregation
When `infer=T` in `makeFS` function, the chisq.test seems to be returning NA's. I suspect this could be because there are counts in a category with a zero probability. Added a couple of lines to check for NA's in this when inferring segregation type and if the case, the segregation type is not inferred.